Gtk-3.2x&4.0: [Budgie-desktop] Fix weird (incorrect) line-indicator colors on checked pseudo-class icontasklists
Those indicators should be drawin as 'on top' layer of elements, but at least
those existed 'beneath' the button node backgrounds. It means filled
alpha-blended backgrounds of button node are overlaid on those indicators.
So draw backgrounds in its image node instead and avoid indicator regions...

It's an upstream issue for me.