Gtk-3.2x&4.0: [GtkCalendar] Shrink day label padding
In GtkCalendar, the 'day' labels (e.g. 1, 2, 3 ...) are aligned with
xalign = center & yalign = top, but on the other hand, the 'week number'
labels are done with xalign = center & yalign = center. That's surely a
technical limitation of Gtk+ theming. There's no way to override the y-align
style-property from user themes.

So this introduces extremely tiny top/bottom padding of labels for keeping
the very same row height between 'day' and 'week number'. And this also means
GtkCalendar labels are too much closely aligned each other in vertical
direction as a side-effect...

This should solve the issue: #9