Gtk-3.24&4.0: [Nautilus] Add a workaround for 'current opened directory' state button styling
In 3.31.90, there's neither .popup/.toggle classes nor :checked pseudo-class,
it means we can't hightlight the last-child (down arrow added) button.

This is the ovbious regression in 3.31 cycle now, so let's try to do hack for
those buttons. But in future, if the upstream implemented some needed states
to those buttons, we will drop this anyway.