Gtk-3.2x&4.0: [Geary] Fill with solid background for the child GtkFrame in left-paned statusbar node
I could not reproduce this issue in various environments though, if
'transparent' GtkFrame was related to this, this commit may or may not help.

GtkFrame is the 1st child container of the parent statusbar node and
covers 100% allocation size of parent, so if the visual glitch happened,
I suspect some unwanted compositing issues might be caused by transparent
GtkFrame instead of statusbar's own background properties...

Related issue: #31.